requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
python-dotenv==1.0.0
mysql-connector-python==8.0.33
scikit-learn==1.3.1
xgboost==1.7.6
pandas==2.0.3
werkzeug==2.3.7
email-validator==2.0.0
